fix: separate commentList and comment component
