+++ frontend/src/scss/page.scss
@@ -554,11 +554,13 @@ PostScreen CSS
   display: flex;
   justify-content: space-between;
   align-items: center;
+  width: 100%;
 }
 
 .commentProfileDiv{
   display: flex;
   height: fit-content;
+  width: 100%;
   justify-content: space-between;
   align-items: center;
   padding-top: 10px;
@@ -574,6 +576,7 @@ PostScreen CSS
   display: flex;
   flex-direction: column;
   margin-left: 5px;
+  width: 100%;
 }
 
 .CommentUserName{
